pyproject.toml
[project]
name = "aiagent"
version = "0.1.0"
description = "Multimodal AI Chef Agent powered by Google Gemini and Custom Search"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.2.0",
    "langchain>=0.0.27",
    "langchain-google-community>=3.0.5",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.8",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
